# docs/api/ui-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Issue Orchestrator UI API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/issues/{issue_number}/timeline-evidence/pin": {
      "put": {
        "summary": "Set the pin state for one exact Timeline run",
        "parameters": [
          {
            "name": "issue_number",
            "in": "path",
            "required": true,
            "schema": {"type": "integer", "minimum": 1}
          },
          {
            "name": "repo_root",
            "in": "query",
            "required": false,
            "description": "Control Center repository scope; omitted on a Repository Engine.",
            "schema": {"type": "string"}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/TimelineEvidencePinRequestPayload"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "The resulting exact-run evidence state",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/TimelineEvidenceStatePayload"}
              }
            }
          },
          "404": {"description": "The exact run or its evidence is unavailable"},
          "409": {"description": "The exact issue/run pair is not eligible for pinning"}
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TimelineEvidencePinRequestPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["run_dir", "pinned"],
        "properties": {
          "run_dir": {"type": "string", "minLength": 1},
          "pinned": {"type": "boolean"}
        }
      },
      "TimelineEvidenceStatePayload": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "run_dir",
          "status",
          "label",
          "available",
          "pinned",
          "archived",
          "expires_at",
          "help_text",
          "unpin_expires_immediately"
        ],
        "properties": {
          "run_dir": {"type": "string"},
          "status": {"type": "string", "enum": ["active", "retained", "pinned", "expired", "missing"]},
          "label": {"type": "string"},
          "available": {"type": "boolean"},
          "pinned": {"type": "boolean"},
          "archived": {"type": "boolean"},
          "expires_at": {"type": ["string", "null"]},
          "help_text": {"type": "string"},
          "unpin_expires_immediately": {"type": "boolean"}
        }
      },
...
    }
  }
}